@@ -4,28 +4,41 @@
     "description": "Adapters, prototypes or functionality that can be potentially incorporated to the C-framework.",
     "keywords": ["framework", "phalcon", "incubator"],
     "homepage": "http://phalconphp.com",
-    "license": "BSD-3",
+    "license": "BSD-3-Clause",
     "authors": [
         {
             "name": "Phalcon Team",
             "email": "team@phalconphp.com"
         }
     ],
+    "support": {
+        "issues": "https://github.com/phalcon/incubator/issues",
+        "source": "https://github.com/phalcon/incubator",
+        "forum": "https://forum.phalconphp.com/"
+    },
     "require": {
-        "php": ">=5.3.21",
-        "ext-phalcon": "~2.0",
-        "swiftmailer/swiftmailer": ">=5.2"
+        "php": ">=5.4",
+        "ext-phalcon": "^2.0",
+        "swiftmailer/swiftmailer": "~5.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.6.*",
-        "squizlabs/php_codesniffer": "2.*",
-        "codeception/codeception": "*",
-        "mockery/mockery": "*",
-        "php-mock/php-mock": "<0.6"
+        "squizlabs/php_codesniffer": "~2.5",
+        "codeception/codeception": "~2.1",
+        "codeception/mockery-module": "~0.2",
+        "codeception/aerospike-module": "~0.1",
+        "codeception/specify": "~0.4",
+        "codeception/verify": "~0.3"
+    },
+    "suggest": {
+        "ext-aerospike": "*",
+        "duncan3dc/fork-helper": "To use extended class to access the beanstalk queue service"
     },
     "autoload": {
-        "psr-0": {
-            "Phalcon": "Library/"
+        "psr-4": { "Phalcon\\": "Library/Phalcon/" }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Phalcon\\Test\\": "tests/unit/"
         }
     }
 }
