@@ -3,7 +3,7 @@
   +------------------------------------------------------------------------+
   | Phalcon Framework                                                      |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2012 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
   | with this package in the file docs/LICENSE.txt.                        |
@@ -16,22 +16,22 @@
   |          Eduar Carvajal <eduar@phalconphp.com>                         |
   +------------------------------------------------------------------------+
 */
+
 namespace Phalcon\Acl\Adapter;
 
 use Phalcon\Acl\Adapter;
-use Phalcon\Acl\AdapterInterface;
 use Phalcon\Acl\Exception;
 use Phalcon\Acl\Resource;
 use Phalcon\Acl;
 use Phalcon\Acl\Role;
+use Phalcon\Acl\RoleInterface;
 
 /**
  * Phalcon\Acl\Adapter\Mongo
  * Manages ACL lists using Mongo Collections
  */
-class Mongo extends Adapter implements AdapterInterface
+class Mongo extends Adapter
 {
-
     /**
      * @var array
      */
@@ -40,7 +40,7 @@ class Mongo extends Adapter implements AdapterInterface
     /**
      * Class constructor.
      *
-     * @param  array                  $options
+     * @param  array $options
      * @throws \Phalcon\Acl\Exception
      */
     public function __construct($options)
@@ -74,6 +74,7 @@ class Mongo extends Adapter implements AdapterInterface
 
     /**
      * {@inheritdoc}
+     *
      * Example:
      * <code>$acl->addRole(new Phalcon\Acl\Role('administrator'), 'consultor');</code>
      * <code>$acl->addRole('administrator', 'consultor');</code>
@@ -81,28 +82,33 @@ class Mongo extends Adapter implements AdapterInterface
      * @param  string  $role
      * @param  array   $accessInherits
      * @return boolean
+     * @throws \Phalcon\Acl\Exception
      */
     public function addRole($role, $accessInherits = null)
     {
-        if (!is_object($role)) {
-            $role = new Role($role);
+        if (is_string($role)) {
+            $role = new Role($role, ucwords($role) . ' Role');
+        }
+
+        if (!$role instanceof RoleInterface) {
+            throw new Exception('Role must be either an string or implement RoleInterface');
         }
 
         $roles = $this->getCollection('roles');
-        $exists = $roles->count(array('name' => $role->getName()));
+        $exists = $roles->count(['name' => $role->getName()]);
 
         if (!$exists) {
-            $roles->insert(array(
+            $roles->insert([
                 'name'        => $role->getName(),
                 'description' => $role->getDescription()
-            ));
+            ]);
 
-            $this->getCollection('accessList')->insert(array(
+            $this->getCollection('accessList')->insert([
                 'roles_name'     => $role->getName(),
                 'resources_name' => '*',
                 'access_name'    => '*',
                 'allowed'        => $this->_defaultAccess
-            ));
+            ]);
         }
 
         if ($accessInherits) {
@@ -115,13 +121,13 @@ class Mongo extends Adapter implements AdapterInterface
     /**
      * {@inheritdoc}
      *
-     * @param  string                 $roleName
-     * @param  string                 $roleToInherit
-     * @throws \Phalcon\Acl\Exception
+     * @param  string $roleName
+     * @param  string $roleToInherit
+     * @throws \BadMethodCallException
      */
     public function addInherit($roleName, $roleToInherit)
     {
-        // TODO: implement later
+        throw new \BadMethodCallException('Not implemented yet.');
     }
 
     /**
@@ -132,7 +138,7 @@ class Mongo extends Adapter implements AdapterInterface
      */
     public function isRole($roleName)
     {
-        return $this->getCollection('roles')->count(array('name' => $roleName)) > 0;
+        return $this->getCollection('roles')->count(['name' => $roleName]) > 0;
     }
 
     /**
@@ -143,11 +149,12 @@ class Mongo extends Adapter implements AdapterInterface
      */
     public function isResource($resourceName)
     {
-        return $this->getCollection('resources')->count(array('name' => $resourceName)) > 0;
+        return $this->getCollection('resources')->count(['name' => $resourceName]) > 0;
     }
 
     /**
      * {@inheritdoc}
+     *
      * Example:
      * <code>
      * //Add a resource to the the list allowing access to an action
@@ -170,12 +177,12 @@ class Mongo extends Adapter implements AdapterInterface
 
         $resources = $this->getCollection('resources');
 
-        $exists = $resources->count(array('name' => $resource->getName()));
+        $exists = $resources->count(['name' => $resource->getName()]);
         if (!$exists) {
-            $resources->insert(array(
+            $resources->insert([
                 'name'        => $resource->getName(),
                 'description' => $resource->getDescription()
-            ));
+            ]);
         }
 
         if ($accessList) {
@@ -188,8 +195,8 @@ class Mongo extends Adapter implements AdapterInterface
     /**
      * {@inheritdoc}
      *
-     * @param  string                 $resourceName
-     * @param  array|string           $accessList
+     * @param  string       $resourceName
+     * @param  array|string $accessList
      * @return boolean
      * @throws \Phalcon\Acl\Exception
      */
@@ -202,19 +209,19 @@ class Mongo extends Adapter implements AdapterInterface
         $resourcesAccesses = $this->getCollection('resourcesAccesses');
 
         if (!is_array($accessList)) {
-            $accessList = array($accessList);
+            $accessList = [$accessList];
         }
 
         foreach ($accessList as $accessName) {
-            $exists = $resourcesAccesses->count(array(
+            $exists = $resourcesAccesses->count([
                 'resources_name' => $resourceName,
                 'access_name'    => $accessName
-            ));
+            ]);
             if (!$exists) {
-                $resourcesAccesses->insert(array(
+                $resourcesAccesses->insert([
                     'resources_name' => $resourceName,
                     'access_name'    => $accessName
-                ));
+                ]);
             }
         }
 
@@ -228,7 +235,7 @@ class Mongo extends Adapter implements AdapterInterface
      */
     public function getResources()
     {
-        $resources = array();
+        $resources = [];
 
         foreach ($this->getCollection('resources')->find() as $row) {
             $resources[] = new Resource($row['name'], $row['description']);
@@ -240,11 +247,11 @@ class Mongo extends Adapter implements AdapterInterface
     /**
      * {@inheritdoc}
      *
-     * @return \Phalcon\Acl\Role[]
+     * @return RoleInterface[]
      */
     public function getRoles()
     {
-        $roles = array();
+        $roles = [];
 
         foreach ($this->getCollection('roles')->find() as $row) {
             $roles[] = new Role($row['name'], $row['description']);
@@ -261,10 +268,12 @@ class Mongo extends Adapter implements AdapterInterface
      */
     public function dropResourceAccess($resourceName, $accessList)
     {
+        throw new \BadMethodCallException('Not implemented yet.');
     }
 
     /**
      * {@inheritdoc}
+     *
      * You can use '*' as wildcard
      * Example:
      * <code>
@@ -289,6 +298,7 @@ class Mongo extends Adapter implements AdapterInterface
 
     /**
      * {@inheritdoc}
+     *
      * You can use '*' as wildcard
      * Example:
      * <code>
@@ -314,6 +324,7 @@ class Mongo extends Adapter implements AdapterInterface
 
     /**
      * {@inheritdoc}
+     *
      * Example:
      * <code>
      * //Does Andres have access to the customers resource to create?
@@ -330,11 +341,11 @@ class Mongo extends Adapter implements AdapterInterface
     public function isAllowed($role, $resource, $access)
     {
         $accessList = $this->getCollection('accessList');
-        $access     = $accessList->findOne(array(
+        $access     = $accessList->findOne([
             'roles_name'     => $role,
             'resources_name' => $resource,
             'access_name'    => $access
-        ));
+        ]);
 
         if (is_array($access)) {
             return (bool) $access['allowed'];
@@ -343,11 +354,12 @@ class Mongo extends Adapter implements AdapterInterface
         /**
          * Check if there is an common rule for that resource
          */
-        $access = $accessList->findOne(array(
+        $access = $accessList->findOne([
             'roles_name'     => $role,
             'resources_name' => $resource,
             'access_name'    => '*'
-        ));
+        ]);
+
         if (is_array($access)) {
             return (bool) $access['allowed'];
         }
@@ -369,10 +381,10 @@ class Mongo extends Adapter implements AdapterInterface
     /**
      * Inserts/Updates a permission in the access list
      *
-     * @param  string                 $roleName
-     * @param  string                 $resourceName
-     * @param  string                 $accessName
-     * @param  integer                $action
+     * @param  string  $roleName
+     * @param  string  $resourceName
+     * @param  string  $accessName
+     * @param  integer $action
      * @return boolean
      * @throws \Phalcon\Acl\Exception
      */
@@ -381,10 +393,11 @@ class Mongo extends Adapter implements AdapterInterface
         /**
          * Check if the access is valid in the resource
          */
-        $exists = $this->getCollection('resourcesAccesses')->count(array(
+        $exists = $this->getCollection('resourcesAccesses')->count([
             'resources_name' => $resourceName,
             'access_name'    => $accessName
-        ));
+        ]);
+
         if (!$exists) {
             throw new Exception(
                 "Access '" . $accessName . "' does not exist in resource '" . $resourceName . "' in ACL"
@@ -393,18 +406,19 @@ class Mongo extends Adapter implements AdapterInterface
 
         $accessList = $this->getCollection('accessList');
 
-        $access = $accessList->findOne(array(
+        $access = $accessList->findOne([
             'roles_name'     => $roleName,
             'resources_name' => $resourceName,
             'access_name'    => $accessName
-        ));
+        ]);
+
         if (!$access) {
-            $accessList->insert(array(
+            $accessList->insert([
                 'roles_name'     => $roleName,
                 'resources_name' => $resourceName,
                 'access_name'    => $accessName,
                 'allowed'        => $action
-            ));
+            ]);
         } else {
             $access['allowed'] = $action;
             $accessList->save($access);
@@ -413,18 +427,19 @@ class Mongo extends Adapter implements AdapterInterface
         /**
          * Update the access '*' in access_list
          */
-        $exists = $accessList->count(array(
+        $exists = $accessList->count([
             'roles_name'     => $roleName,
             'resources_name' => $resourceName,
             'access_name'    => '*'
-        ));
+        ]);
+
         if (!$exists) {
-            $accessList->insert(array(
+            $accessList->insert([
                 'roles_name'     => $roleName,
                 'resources_name' => $resourceName,
                 'access_name'    => '*',
                 'allowed'        => $this->_defaultAccess
-            ));
+            ]);
         }
 
         return true;
@@ -433,10 +448,10 @@ class Mongo extends Adapter implements AdapterInterface
     /**
      * Inserts/Updates a permission in the access list
      *
-     * @param  string                 $roleName
-     * @param  string                 $resourceName
-     * @param  string                 $access
-     * @param  integer                $action
+     * @param  string  $roleName
+     * @param  string  $resourceName
+     * @param  string  $access
+     * @param  integer $action
      * @throws \Phalcon\Acl\Exception
      */
     protected function allowOrDeny($roleName, $resourceName, $access, $action)
@@ -446,7 +461,7 @@ class Mongo extends Adapter implements AdapterInterface
         }
 
         if (!is_array($access)) {
-            $access = array($access);
+            $access = [$access];
         }
 
         foreach ($access as $accessName) {
