@@ -3,7 +3,7 @@
   +------------------------------------------------------------------------+
   | Phalcon Framework                                                      |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2012 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
   | with this package in the file docs/LICENSE.txt.                        |
@@ -17,22 +17,30 @@
   |          Nikita Vershinin <endeveit@gmail.com>                         |
   +------------------------------------------------------------------------+
 */
+
 namespace Phalcon\Cache\Backend;
 
 use Phalcon\Cache\Exception;
 use Phalcon\Cache\FrontendInterface;
 use Phalcon\Db;
 use Phalcon\Db\AdapterInterface as DbAdapterInterface;
+use Phalcon\Cache\Backend;
+use Phalcon\Cache\BackendInterface;
 
 /**
  * Phalcon\Cache\Backend\Database
+ *
  * This backend uses a database as cache backend
+ *
+ * @package Phalcon\Cache\Backend
+ * @property \Phalcon\Cache\FrontendInterface _frontend
  */
-class Database extends Prefixable
+class Database extends Backend implements BackendInterface
 {
+    use Prefixable;
 
     /**
-     * @var \Phalcon\Db\AdapterInterface
+     * @var DbAdapterInterface
      */
     protected $db = null;
 
@@ -42,28 +50,28 @@ class Database extends Prefixable
     protected $table = null;
 
     /**
-     * Class constructor.
+     * {@inheritdoc}
      *
-     * @param  \Phalcon\Cache\FrontendInterface $frontend
-     * @param  array                            $options
-     * @throws \Phalcon\Cache\Exception
+     * @param  FrontendInterface $frontend
+     * @param  array             $options
+     * @throws Exception
      */
-    public function __construct(FrontendInterface $frontend, $options = array())
+    public function __construct(FrontendInterface $frontend, array $options)
     {
-        if (!isset($options['db'])) {
-            throw new Exception("Parameter 'db' is required");
+        if (!isset($options['db']) || !$options['db'] instanceof DbAdapterInterface) {
+            throw new Exception(
+                'Parameter "db" is required and it must be an instance of Phalcon\Acl\AdapterInterface'
+            );
         }
 
-        if (!($options['db'] instanceof DbAdapterInterface)) {
-            throw new Exception("Parameter 'db' must implement Phalcon\\Db\\AdapterInterface");
-        }
-
-        if (!isset($options['table'])) {
-            throw new Exception("Parameter 'table' is required");
+        if (!isset($options['table']) || empty($options['table']) || !is_string($options['table'])) {
+            throw new Exception("Parameter 'table' is required and it must be a non empty string");
         }
 
         $this->db    = $options['db'];
-        $this->table = $options['table'];
+        $this->table = $this->db->escapeIdentifier($options['table']);
+
+        unset($options['db'], $options['table']);
 
         parent::__construct($frontend, $options);
     }
@@ -78,8 +86,8 @@ class Database extends Prefixable
     public function get($keyName, $lifetime = null)
     {
         $prefixedKey    = $this->getPrefixedIdentifier($keyName);
-        $sql            = "SELECT data, lifetime FROM " . $this->table . " WHERE key_name = ?";
-        $cache          = $this->db->fetchOne($sql, Db::FETCH_ASSOC, array($prefixedKey));
+        $sql            = "SELECT data, lifetime FROM {$this->table} WHERE key_name = ?";
+        $cache          = $this->db->fetchOne($sql, Db::FETCH_ASSOC, [$prefixedKey]);
         $this->_lastKey = $prefixedKey;
 
         if (!$cache) {
@@ -91,7 +99,7 @@ class Database extends Prefixable
 
         // Remove the cache if expired
         if ($cache['lifetime'] < time()) {
-            $this->db->execute("DELETE FROM " . $this->table . " WHERE key_name = ?", array($prefixedKey));
+            $this->db->execute("DELETE FROM {$this->table} WHERE key_name = ?", [$prefixedKey]);
 
             return null;
         }
@@ -102,10 +110,10 @@ class Database extends Prefixable
     /**
      * {@inheritdoc}
      *
-     * @param  string                   $keyName
-     * @param  string                   $content
-     * @param  integer                  $lifetime
-     * @param  boolean                  $stopBuffer
+     * @param  string $keyName
+     * @param  string $content
+     * @param  int    $lifetime
+     * @param  bool   $stopBuffer
      * @throws \Phalcon\Cache\Exception
      */
     public function save($keyName = null, $content = null, $lifetime = null, $stopBuffer = true)
@@ -136,23 +144,23 @@ class Database extends Prefixable
         $lifetime = time() + $lifetime;
 
         // Check if the cache already exist
-        $sql   = "SELECT data, lifetime FROM " . $this->table . " WHERE key_name = ?";
+        $sql   = "SELECT data, lifetime FROM {$this->table} WHERE key_name = ?";
         $cache = $this->db->fetchOne($sql, Db::FETCH_ASSOC, array($prefixedKey));
 
         if (!$cache) {
-            $this->db->execute("INSERT INTO " . $this->table . " VALUES (?, ?, ?)", array(
+            $this->db->execute("INSERT INTO {$this->table} VALUES (?, ?, ?)", [
                 $prefixedKey,
                 $frontend->beforeStore($cachedContent),
                 $lifetime
-            ));
+            ]);
         } else {
             $this->db->execute(
-                "UPDATE " . $this->table . " SET data = ?, lifetime = ? WHERE key_name = ?",
-                array(
+                "UPDATE {$this->table} SET data = ?, lifetime = ? WHERE key_name = ?",
+                [
                     $frontend->beforeStore($cachedContent),
                     $lifetime,
                     $prefixedKey
-                )
+                ]
             );
         }
 
@@ -171,19 +179,19 @@ class Database extends Prefixable
      * {@inheritdoc}
      *
      * @param  string  $keyName
-     * @return boolean
+     * @return bool
      */
     public function delete($keyName)
     {
         $prefixedKey = $this->getPrefixedIdentifier($keyName);
-        $sql         = "SELECT COUNT(*) AS rowcount FROM " . $this->table . " WHERE key_name = ?";
-        $row         = $this->db->fetchOne($sql, Db::FETCH_ASSOC, array($prefixedKey));
+        $sql         = "SELECT COUNT(*) AS rowcount FROM {$this->table} WHERE key_name = ?";
+        $row         = $this->db->fetchOne($sql, Db::FETCH_ASSOC, [$prefixedKey]);
 
         if (!$row['rowcount']) {
             return false;
         }
 
-        return $this->db->execute("DELETE FROM " . $this->table . " WHERE key_name = ?", array($prefixedKey));
+        return $this->db->execute("DELETE FROM {$this->table} WHERE key_name = ?", [$prefixedKey]);
     }
 
     /**
@@ -201,16 +209,16 @@ class Database extends Prefixable
         }
 
         if (!empty($prefix)) {
-            $sql = "SELECT key_name FROM " . $this->table . " WHERE key_name LIKE ? ORDER BY lifetime";
-            $rs  = $this->db->query($sql, array($prefix . '%'));
+            $sql = "SELECT key_name FROM {$this->table} WHERE key_name LIKE ? ORDER BY lifetime";
+            $rs  = $this->db->query($sql, [$prefix . '%']);
         } else {
-            $sql = "SELECT key_name FROM " . $this->table . " ORDER BY lifetime";
+            $sql = "SELECT key_name FROM {$this->table} ORDER BY lifetime";
             $rs  = $this->db->query($sql);
         }
 
         $rs->setFetchMode(Db::FETCH_ASSOC);
 
-        $keys = array();
+        $keys = [];
 
         while ($row = $rs->fetch()) {
             $keys[] = !empty($prefix) ? str_replace($prefix, '', $row['key_name']) : $row['key_name'];
@@ -224,13 +232,13 @@ class Database extends Prefixable
      *
      * @param  string  $keyName
      * @param  string  $lifetime
-     * @return boolean
+     * @return bool
      */
     public function exists($keyName = null, $lifetime = null)
     {
         $prefixedKey = $this->getPrefixedIdentifier($keyName);
-        $sql         = "SELECT lifetime FROM " . $this->table . " WHERE key_name = ?";
-        $cache       = $this->db->fetchOne($sql, Db::FETCH_ASSOC, array($prefixedKey));
+        $sql         = "SELECT lifetime FROM {$this->table} WHERE key_name = ?";
+        $cache       = $this->db->fetchOne($sql, Db::FETCH_ASSOC, [$prefixedKey]);
 
         if (!$cache) {
             return false;
@@ -238,7 +246,7 @@ class Database extends Prefixable
 
         // Remove the cache if expired
         if ($cache['lifetime'] < time()) {
-            $this->db->execute("DELETE FROM " . $this->table . " WHERE key_name = ?", array($prefixedKey));
+            $this->db->execute("DELETE FROM {$this->table} WHERE key_name = ?", [$prefixedKey]);
 
             return false;
         }
@@ -249,11 +257,11 @@ class Database extends Prefixable
     /**
      * {@inheritdoc}
      *
-     * @return boolean
+     * @return bool
      */
     public function flush()
     {
-        $this->db->execute('DELETE FROM ' . $this->table);
+        $this->db->execute("DELETE FROM {$this->table}");
 
         return true;
     }
