@@ -3,7 +3,7 @@
   +------------------------------------------------------------------------+
   | Phalcon Framework                                                      |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2012 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
   | with this package in the file docs/LICENSE.txt.                        |
@@ -19,9 +19,20 @@
 */
 namespace Phalcon\Error;
 
+/**
+ * Class Error
+ * @package Phalcon\Error
+ *
+ * @method int type()
+ * @method string message()
+ * @method string file()
+ * @method string line()
+ * @method \Exception exception()
+ * @method bool isException()
+ * @method bool isError()
+ */
 class Error
 {
-
     /**
      * @var array
      */
@@ -32,9 +43,9 @@ class Error
      *
      * @param array $options
      */
-    public function __construct(array $options = array())
+    public function __construct(array $options = [])
     {
-        $defaults = array(
+        $defaults = [
             'type'        => -1,
             'message'     => 'No error message',
             'file'        => '',
@@ -42,7 +53,7 @@ class Error
             'exception'   => null,
             'isException' => false,
             'isError'     => false,
-        );
+        ];
 
         $options = array_merge($defaults, $options);
 
