@@ -1,53 +1,111 @@
 <?php
+/*
+  +------------------------------------------------------------------------+
+  | Phalcon Framework                                                      |
+  +------------------------------------------------------------------------+
+  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
+  +------------------------------------------------------------------------+
+  | This source file is subject to the New BSD License that is bundled     |
+  | with this package in the file docs/LICENSE.txt.                        |
+  |                                                                        |
+  | If you did not receive a copy of the license and are unable to         |
+  | obtain it through the world-wide-web, please send an email             |
+  | to license@phalconphp.com so we can send you a copy immediately.       |
+  +------------------------------------------------------------------------+
+  | Authors: Andres Gutierrez <andres@phalconphp.com>                      |
+  +------------------------------------------------------------------------+
+*/
+
 namespace Phalcon\Logger\Adapter;
 
+use Phalcon\Db\Column;
 use Phalcon\Logger\Exception;
+use Phalcon\Logger\Formatter\Line as LineFormatter;
+use Phalcon\Logger\Adapter as LoggerAdapter;
+use Phalcon\Logger\AdapterInterface;
+use Phalcon\Db\AdapterInterface as DbAdapterInterface;
 
 /**
- * Phalcon\Logger\Adapter\Database
+ * Database Logger
+ *
  * Adapter to store logs in a database table
+ *
+ * @package Phalcon\Logger\Adapter
  */
-class Database extends \Phalcon\Logger\Adapter implements \Phalcon\Logger\AdapterInterface
+class Database extends LoggerAdapter implements AdapterInterface
 {
-
     /**
      * Name
+     * @var string
      */
-    protected $name;
+    protected $name = 'phalcon';
 
     /**
      * Adapter options
+     * @var array
      */
-    protected $options;
+    protected $options = [];
+
+    /**
+     * @var \Phalcon\Db\AdapterInterface
+     */
+    protected $db;
 
     /**
      * Class constructor.
      *
-     * @param  string                    $name
-     * @param  array                     $options
+     * @param  string $name
+     * @param  array  $options
      * @throws \Phalcon\Logger\Exception
      */
-    public function __construct($name, $options = array())
+    public function __construct($name = 'phalcon', array $options = [])
     {
         if (!isset($options['db'])) {
             throw new Exception("Parameter 'db' is required");
+        }
+
+        if (!$options['db'] instanceof DbAdapterInterface) {
+            throw new Exception("Parameter 'db' must be object and implement AdapterInterface");
         }
 
         if (!isset($options['table'])) {
             throw new Exception("Parameter 'table' is required");
         }
 
-        $this->name = $name;
+        $this->db = $options['db'];
+
+        if ($name) {
+            $this->name = $name;
+        }
+
         $this->options = $options;
+    }
+
+    /**
+     * Sets database connection
+     *
+     * @param AdapterInterface $db
+     * @return $this
+     */
+    public function setDb(AdapterInterface $db)
+    {
+        $this->db = $db;
+
+        return $this;
     }
 
     /**
      * {@inheritdoc}
      *
-     * @return \Phalcon\Logger\Formatter\Line
+     * @return \Phalcon\Logger\FormatterInterface
      */
     public function getFormatter()
     {
+        if (!is_object($this->_formatter)) {
+            $this->_formatter = new LineFormatter();
+        }
+
+        return $this->_formatter;
     }
 
     /**
@@ -57,12 +115,14 @@ class Database extends \Phalcon\Logger\Adapter implements \Phalcon\Logger\Adapte
      * @param integer $type
      * @param integer $time
      * @param array   $context
+     * @return bool
      */
-    public function logInternal($message, $type, $time, $context = array())
+    public function logInternal($message, $type, $time, $context = [])
     {
-        return $this->options['db']->execute(
+        return $this->db->execute(
             'INSERT INTO ' . $this->options['table'] . ' VALUES (null, ?, ?, ?, ?)',
-            array($this->name, $type, $message, $time)
+            [$this->name, $type, $message, $time],
+            [Column::BIND_PARAM_STR, Column::BIND_PARAM_INT, Column::BIND_PARAM_STR, Column::BIND_PARAM_INT]
         );
     }
 
@@ -73,6 +133,49 @@ class Database extends \Phalcon\Logger\Adapter implements \Phalcon\Logger\Adapte
      */
     public function close()
     {
-        $this->options['db']->close();
+        if ($this->db->isUnderTransaction()) {
+            $this->db->commit();
+        }
+
+        $this->db->close();
+
+        return true;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @return $this
+     */
+    public function begin()
+    {
+        $this->db->begin();
+
+        return $this;
+    }
+
+    /**
+     * Commit transaction
+     *
+     * @return $this
+     */
+    public function commit()
+    {
+        $this->db->commit();
+
+        return $this;
+    }
+
+    /**
+     * Rollback transaction
+     * (happens automatically if commit never reached)
+     *
+     * @return $this
+     */
+    public function rollback()
+    {
+        $this->db->rollback();
+
+        return $this;
     }
 }
