@@ -4,7 +4,7 @@
  +------------------------------------------------------------------------+
  | Phalcon Framework                                                      |
  +------------------------------------------------------------------------+
- | Copyright (c) 2011-2015 Phalcon Team (http://www.phalconphp.com)       |
+ | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
  +------------------------------------------------------------------------+
  | This source file is subject to the New BSD License that is bundled     |
  | with this package in the file docs/LICENSE.txt.                        |
@@ -44,12 +44,12 @@ class Extended extends ConsoleApp
      */
     public function handle(array $arguments = null)
     {
-        if (isset($arguments['task']) && in_array($arguments['task'], array('-h','--help','help'))) {
+        if (isset($arguments['task']) && in_array($arguments['task'], array('-h', '--help', 'help'))) {
             $this->setTasksDir();
             $this->createHelp();
             $this->showHelp();
             return;
-        } elseif (isset($arguments['action']) && in_array($arguments['action'], array('-h','--help','help'))) {
+        } elseif (isset($arguments['action']) && in_array($arguments['action'], array('-h', '--help', 'help'))) {
             $this->setTasksDir();
             $this->createHelp();
             $this->showTaskHelp($arguments['task']);
@@ -93,9 +93,9 @@ class Extended extends ConsoleApp
         }
 
         foreach ($scannedTasksDir as $taskFile) {
-            $taskClass = ($namespace ? $namespace . '\\' : '') . str_replace('.php', '', $taskFile);
-            $taskName  = strtolower(str_replace('Task', '', $taskClass));
-            $taskName  = trim($taskName, '\\');
+            $taskFileInfo = pathinfo($taskFile);
+            $taskClass = ($namespace ? $namespace . '\\' : '') . $taskFileInfo["filename"];
+            $taskName  = strtolower(str_replace('Task', '', $taskFileInfo["filename"]));
 
             $this->documentation[$taskName] = array('description'=>array(''), 'actions'=>array());
 
