@@ -4,7 +4,7 @@
   +------------------------------------------------------------------------+
   | Phalcon Framework                                                      |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2014 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
   | with this package in the file docs/LICENSE.txt.                        |
@@ -25,43 +25,51 @@ namespace Phalcon\Db\Dialect;
  *
  * Every query executed via this adapter is automatically cached
  */
-class MysqlExtended extends \Phalcon\Db\Dialect\Mysql
+class MysqlExtended extends Mysql
 {
-
     /**
      * Transforms an intermediate representation for a expression into a database system valid expression
      *
-     * @param array  expression
-     * @param string escapeChar
+     * @param array  $expression
+     * @param string $escapeChar
      *
      * @return string
+     * @throws \Exception
      */
-    public function getSqlExpression($expression, $escapeChar = null)
+    public function getSqlExpression(array $expression, $escapeChar = null, $bindCounts = null)
     {
         if ($expression["type"] == 'functionCall') {
             switch ($expression["name"]) {
-
                 case 'DATE_INTERVAL':
-
                     if (count($expression["arguments"]) != 2) {
                         throw new \Exception('DATE_INTERVAL requires 2 parameters');
                     }
 
                     switch ($expression["arguments"][1]['value']) {
-
+                        case "'MICROSECOND'":
+                            return 'INTERVAL ' . $this->getSqlExpression($expression["arguments"][0]) . ' MICROSECOND';
+                        case "'SECOND'":
+                            return 'INTERVAL ' . $this->getSqlExpression($expression["arguments"][0]) . ' SECOND';
+                        case "'MINUTE'":
+                            return 'INTERVAL ' . $this->getSqlExpression($expression["arguments"][0]) . ' MINUTE';
+                        case "'HOUR'":
+                            return 'INTERVAL ' . $this->getSqlExpression($expression["arguments"][0]) . ' HOUR';
                         case "'DAY'":
                             return 'INTERVAL ' . $this->getSqlExpression($expression["arguments"][0]) . ' DAY';
-
+                        case "'WEEK'":
+                            return 'INTERVAL ' . $this->getSqlExpression($expression["arguments"][0]) . ' WEEK';
                         case "'MONTH'":
                             return 'INTERVAL ' . $this->getSqlExpression($expression["arguments"][0]) . ' MONTH';
-
+                        case "'QUARTER'":
+                            return 'INTERVAL ' . $this->getSqlExpression($expression["arguments"][0]) . ' QUARTER';
                         case "'YEAR'":
                             return 'INTERVAL ' . $this->getSqlExpression($expression["arguments"][0]) . ' YEAR';
+                        default:
+                            throw new \Exception('DATE_INTERVAL unit is not supported');
                     }
                     break;
 
                 case 'FULLTEXT_MATCH':
-
                     if (count($expression["arguments"]) < 2) {
                         throw new \Exception('FULLTEXT_MATCH requires 2 parameters');
                     }
@@ -76,7 +84,6 @@ class MysqlExtended extends \Phalcon\Db\Dialect\Mysql
                     $this->getSqlExpression($expression["arguments"][$length]) . ')';
 
                 case 'FULLTEXT_MATCH_BMODE':
-
                     if (count($expression["arguments"]) < 2) {
                         throw new \Exception('FULLTEXT_MATCH requires 2 parameters');
                     }
