@@ -12,7 +12,6 @@ use Phalcon\Mvc\ViewBaseInterface;
  */
 class Smarty extends Engine implements EngineInterface
 {
-
     /**
      * @var \Smarty
      */
@@ -50,7 +49,11 @@ class Smarty extends Engine implements EngineInterface
             $params['content'] = $this->_view->getContent();
         }
         foreach ($params as $key => $value) {
-            $this->smarty->assign($key, $value);
+            if (isset($params['_' . $key]) && $params['_' . $key] === true) {
+                $this->smarty->assign($key, $value, true);
+            } else {
+                $this->smarty->assign($key, $value);
+            }
         }
 
         $content = $this->smarty->fetch($path);
