@@ -1,95 +1,93 @@
 <?php
-/**
- * UnitTestCase.php
- * Phalcon_Test_UnitTestCase
- * Unit Test Helper
- * PhalconPHP Framework
- *
- * @copyright (c) 2011-2015 Phalcon Team
- * @link          http://www.phalconphp.com
- * @author        Andres Gutierrez <andres@phalconphp.com>
- * @author        Nikolaos Dimopoulos <nikos@phalconphp.com>
- * @author        Stephen Hoogendijk <hoogendijk09@gmail.com>
- *                The contents of this file are subject to the New BSD License that is
- *                bundled with this package in the file docs/LICENSE.txt
- *                If you did not receive a copy of the license and are unable to obtain it
- *                through the world-wide-web, please send an email to license@phalconphp.com
- *                so that we can send you a copy immediately.
- */
+/*
+  +------------------------------------------------------------------------+
+  | Phalcon Framework                                                      |
+  +------------------------------------------------------------------------+
+  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
+  +------------------------------------------------------------------------+
+  | This source file is subject to the New BSD License that is bundled     |
+  | with this package in the file docs/LICENSE.txt.                        |
+  |                                                                        |
+  | If you did not receive a copy of the license and are unable to         |
+  | obtain it through the world-wide-web, please send an email             |
+  | to license@phalconphp.com so we can send you a copy immediately.       |
+  +------------------------------------------------------------------------+
+  | Authors: Stephen Hoogendijk <stephen@tca0.nl>                          |
+  +------------------------------------------------------------------------+
+*/
+
 namespace Phalcon\Test;
 
+use Phalcon\Di\InjectionAwareInterface;
+use PHPUnit_Framework_TestCase as TestCase;
 use Phalcon\Config;
-use Phalcon\DI\FactoryDefault;
-use Phalcon\DI;
+use Phalcon\Di\FactoryDefault;
+use Phalcon\Di;
 use Phalcon\DiInterface;
 use Phalcon\Mvc\Url;
+use Phalcon\Escaper;
 
 /**
  * Class UnitTestCase
  *
  * @package Phalcon\Test
  */
-abstract class UnitTestCase extends \PHPUnit_Framework_TestCase
+abstract class UnitTestCase extends TestCase implements InjectionAwareInterface
 {
-
     /**
      * Holds the configuration variables and other stuff
      * I can use the DI container but for tests like the Translate
      * we do not need the overhead
      *
-     * @var array
+     * @var Config|null
      */
-    protected $config = array();
+    protected $config;
 
     /**
-     * @var \Phalcon\DiInterface
+     * @var DiInterface
      */
     protected $di;
 
     /**
-     * Sets the test up by loading the DI container and other stuff
-     *
-     * @author Nikos Dimopoulos <nikos@phalconphp.com>
-     * @since  2012-09-30
-     * @param  \Phalcon\DiInterface $di
-     * @param  \Phalcon\Config      $config
-     * @return void
+     * This method is called before a test is executed.
      */
-    protected function setUp(DiInterface $di = null, Config $config = null)
+    protected function setUp()
     {
         $this->checkExtension('phalcon');
 
-        if (!is_null($config)) {
-            $this->config = $config;
-        }
+        // Reset the DI container
+        Di::reset();
 
-        if (is_null($di)) {
-            // Reset the DI container
-            DI::reset();
+        // Instantiate a new DI container
+        $di = new Di();
 
-            // Instantiate a new DI container
-            $di = new FactoryDefault();
+        // Set the URL
+        $di->set(
+            'url',
+            function () {
+                $url = new Url();
+                $url->setBaseUri('/');
 
-            // Set the URL
-            $di->set(
-                'url',
-                function () {
-                    $url = new Url();
-                    $url->setBaseUri('/');
+                return $url;
+            }
+        );
 
-                    return $url;
-                }
-            );
-
-            $di->set(
-                'escaper',
-                function () {
-                    return new \Phalcon\Escaper();
-                }
-            );
-        }
+        $di->set(
+            'escaper',
+            function () {
+                return new Escaper();
+            }
+        );
 
         $this->di = $di;
+    }
+
+    protected function tearDown()
+    {
+        $di = $this->getDI();
+        $di::reset();
+
+        parent::tearDown();
     }
 
     /**
@@ -119,8 +117,6 @@ abstract class UnitTestCase extends \PHPUnit_Framework_TestCase
     /**
      * Returns a unique file name
      *
-     * @author Nikos Dimopoulos <nikos@phalconphp.com>
-     * @since  2012-09-30
      * @param  string $prefix A prefix for the file
      * @param  string $suffix A suffix for the file
      * @return string
@@ -136,10 +132,8 @@ abstract class UnitTestCase extends \PHPUnit_Framework_TestCase
     /**
      * Removes a file from the system
      *
-     * @author Nikos Dimopoulos <nikos@phalconphp.com>
-     * @since  2012-09-30
-     * @param $path
-     * @param $fileName
+     * @param string $path
+     * @param string $fileName
      */
     protected function cleanFile($path, $fileName)
     {
@@ -154,17 +148,58 @@ abstract class UnitTestCase extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @return \Phalcon\DiInterface
+     * Sets the Config object.
+     *
+     * @param Config $config
+     * @return $this
      */
-    protected function getDI()
+    public function setConfig(Config $config)
     {
-        return $this->di;
+        $this->config = $config;
+
+        return $this;
     }
 
-    protected function tearDown()
+    /**
+     * Returns the Config object if any.
+     *
+     * @return null|Config
+     */
+    public function getConfig()
     {
-        $di = $this->getDI();
-        $di::reset();
-        parent::tearDown();
+        if (!$this->config instanceof Config && $this->getDI()->has('config')) {
+            return $this->getDI()->has('config');
+        }
+
+        return $this->config;
+    }
+
+    /**
+     * Sets the Dependency Injector.
+     *
+     * @see    Injectable::setDI
+     * @param  DiInterface $di
+     * @return $this
+     */
+    public function setDI(DiInterface $di)
+    {
+        $this->di = $di;
+
+        return $this;
+    }
+
+    /**
+     * Returns the internal Dependency Injector.
+     *
+     * @see    Injectable::getDI
+     * @return DiInterface
+     */
+    public function getDI()
+    {
+        if (!$this->di instanceof DiInterface) {
+            return Di::getDefault();
+        }
+
+        return $this->di;
     }
 }
