@@ -3,7 +3,7 @@
   +------------------------------------------------------------------------+
   | Phalcon Framework                                                      |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2012 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
   | with this package in the file docs/LICENSE.txt.                        |
@@ -16,27 +16,40 @@
   |          Eduar Carvajal <eduar@phalconphp.com>                         |
   +------------------------------------------------------------------------+
 */
+
 namespace Phalcon\Translate\Adapter;
 
+use Phalcon\Db;
 use Phalcon\Translate\Adapter;
 use Phalcon\Translate\AdapterInterface;
 use Phalcon\Translate\Exception;
 
-class Database extends Base implements AdapterInterface
+class Database extends Base implements AdapterInterface, \ArrayAccess
 {
-
     /**
      * @var array
      */
     protected $options;
 
     /**
+     * Statement for Exist
+     * @var array
+     */
+    protected $stmtExists;
+
+    /**
+     * Statement for Read
+     * @var array
+     */
+    protected $stmtSelect;
+
+    /**
      * Class constructor.
      *
-     * @param  array                        $options
+     * @param  array $options
      * @throws \Phalcon\Translate\Exception
      */
-    public function __construct($options)
+    public function __construct(array $options)
     {
         if (!isset($options['db'])) {
             throw new Exception("Parameter 'db' is required");
@@ -50,53 +63,184 @@ class Database extends Base implements AdapterInterface
             throw new Exception("Parameter 'language' is required");
         }
 
+        $this->stmtSelect = sprintf(
+            'SELECT value FROM %s WHERE language = :language AND key_name = :key_name',
+            $options['table']
+        );
+
+        $this->stmtExists = sprintf(
+            'SELECT COUNT(*) AS `count` FROM %s WHERE language = :language AND key_name = :key_name',
+            $options['table']
+        );
+
         $this->options = $options;
     }
 
     /**
      * {@inheritdoc}
      *
-     * @param  string $index
+     * @param  string $translateKey
      * @param  array  $placeholders
      * @return string
      */
-    public function query($index, $placeholders = null)
+    public function query($translateKey, $placeholders = null)
     {
         $options = $this->options;
 
         $translation = $options['db']->fetchOne(
-            sprintf(
-                "SELECT value FROM %s WHERE language = '%s' AND key_name = ?",
-                $options['table'],
-                $options['language']
-            ),
-            null,
-            array($index)
+            $this->stmtSelect,
+            Db::FETCH_ASSOC,
+            ['language' => $options['language'], 'key_name' => $translateKey]
         );
 
-        if (!$translation) {
-            return $index;
+        $value = empty($translation['value']) ? $translateKey : $translation['value'];
+
+        if (is_array($placeholders)) {
+            foreach ($placeholders as $k => $v) {
+                $value = str_replace('%' . $k . '%', $v, $value);
+            }
         }
 
-        return self::setPlaceholders($translation['value'], $placeholders);
+        return $value;
+    }
+
+    /**
+     * Returns the translation string of the given key
+     *
+     * @param  string $translateKey
+     * @param  array  $placeholders
+     * @return string
+     */
+    // @codingStandardsIgnoreStart
+    public function _($translateKey, $placeholders = null)
+    {
+        return $this->query($translateKey, $placeholders);
+    }
+    // @codingStandardsIgnoreEnd
+
+    /**
+     * {@inheritdoc}
+     *
+     * @param  string  $translateKey
+     * @return boolean
+     */
+    public function exists($translateKey)
+    {
+        $options = $this->options;
+
+        $result = $options['db']->fetchOne(
+            $this->stmtExists,
+            Db::FETCH_ASSOC,
+            ['language' => $options['language'], 'key_name' => $translateKey]
+        );
+
+        return !empty($result['count']);
+    }
+
+    /**
+     * Adds a translation for given key (No existence check!)
+     *
+     * @param  string  $translateKey
+     * @param  string  $message
+     * @return boolean
+     */
+    public function add($translateKey, $message)
+    {
+        $options = $this->options;
+        $data = ['language' => $options['language'], 'key_name' => $translateKey, 'value' => $message];
+
+        return $options['db']->insert($options['table'], array_values($data), array_keys($data));
+    }
+
+    /**
+     * Update a translation for given key (No existence check!)
+     *
+     * @param  string  $translateKey
+     * @param  string  $message
+     * @return boolean
+     */
+    public function update($translateKey, $message)
+    {
+        $options = $this->options;
+
+        return $options['db']->update($options['table'], ['value'], [$message], [
+            'conditions' => 'key_name = ? AND language = ?',
+            'bind' => ['key' => $translateKey, 'lang' => $options['language']]
+        ]);
+    }
+
+    /**
+     * Deletes a translation for given key (No existence check!)
+     *
+     * @param  string  $translateKey
+     * @return boolean
+     */
+    public function delete($translateKey)
+    {
+        $options = $this->options;
+
+        return $options['db']->delete(
+            $options['table'],
+            'key_name = :key AND language = :lang',
+            ['key' => $translateKey, 'lang' => $options['language']]
+        );
+    }
+
+    /**
+     * Sets (insert or updates) a translation for given key
+     *
+     * @param  string  $translateKey
+     * @param  string  $message
+     * @return boolean
+     */
+    public function set($translateKey, $message)
+    {
+        return $this->exists($translateKey) ?
+            $this->update($translateKey, $message) : $this->add($translateKey, $message);
     }
 
     /**
      * {@inheritdoc}
      *
-     * @param  string  $index
-     * @return boolean
+     * @param  string $translateKey
+     * @return string
      */
-    public function exists($index)
+    public function offsetExists($translateKey)
     {
-        $options = $this->options;
+        return $this->exists($translateKey);
+    }
 
-        $exists = $options['db']->fetchOne(
-            "SELECT COUNT(*) FROM " . $options['table'] . " WHERE key_name = ?0",
-            null,
-            array($index)
-        );
+    /**
+     * {@inheritdoc}
+     *
+     * @param  string $translateKey
+     * @param  string $message
+     * @return string
+     */
+    public function offsetSet($translateKey, $message)
+    {
+        return $this->update($translateKey, $message);
+    }
 
-        return $exists[0] > 0;
+    /**
+     * {@inheritdoc}
+     *
+     * @param string $translateKey
+     * @return string
+     */
+    public function offsetGet($translateKey)
+    {
+        return $this->query($translateKey);
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @param  string $translateKey
+     * @return string
+     */
+    public function offsetUnset($translateKey)
+    {
+        return $this->delete($translateKey);
     }
 }
